Ensure the value of `SSL_CERT_FILE` is converted to a string before setting
As is, `Rails.root.join('tmp', 'cacert.crt')` will return a `Pathname` object [1].
In order to set `ENV['SSL_CERT_FILE']`, the value must be a string [2], meaning
we need to call `.to_s` [3] on the output of the `Pathname` in order for it to set
correctly, without erroring: `TypeError: no implicit conversion of Pathname into String`.

The initial PR to introduce this change [4] was extracted from a compliance PR [5]
where the value of `ENV['SSL_CERT_FILE']` was being explicitly set to a string.

[1] https://ruby-doc.org/3.1.2/exts/pathname/Pathname.html
[2] https://ruby-doc.org/3.1.2/ENV.html
[3] https://ruby-doc.org/3.1.2/exts/pathname/Pathname.html#method-i-to_s
[4] https://github.com/RedHatInsights/clowder-common-ruby/pull/32
[5] https://github.com/RedHatInsights/compliance-backend/pull/1664/files#diff-274fbb104eb7e6e64d378ecb4f260a28440db9fc32916c19337dae8aa46cf85aL14